+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/ExtensionConfigurator.java
@@ -1,6 +1,7 @@
 package ru.yandex.money.gradle.plugins.gradleproject;
 
 import org.gradle.api.Project;
+import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestExtension;
 import ru.yandex.money.gradle.plugins.gradleproject.git.GitManager;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.EmailConnectionExtension;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.GitConnectionExtension;
@@ -24,6 +25,7 @@ public class ExtensionConfigurator {
     static void configure(Project project) {
         configureGitExpiredBranchesPlugin(project);
         configureReleasePlugin(project);
+        configureArchitectureTestPlugin(project);
     }
 
     private static void configureReleasePlugin(Project project) {
@@ -53,5 +55,12 @@ public class ExtensionConfigurator {
                 project.getExtensions().findByType(GitConnectionExtension.class);
 
         gitConnectionExtension.setPathToGitPrivateSshKey(System.getenv("GIT_PRIVATE_SSH_KEY_PATH"));
+    }
+
+    private static void configureArchitectureTestPlugin(Project project) {
+        ArchitectureTestExtension architectureTestExtension = project.getExtensions().getByType(ArchitectureTestExtension.class);
+        architectureTestExtension.getInclude().addAll(Arrays.asList(
+                "check_unique_enums_codes",
+                "check_unique_enums_secondary_codes"));
     }
 }