+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/GradleProjectPlugin.java
@@ -5,10 +5,12 @@ import org.gradle.api.Project;
 import org.gradle.api.publish.maven.plugins.MavenPublishPlugin;
 import org.gradle.plugin.devel.plugins.JavaGradlePluginPlugin;
 import org.gradle.util.VersionNumber;
+import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestPlugin;
 import ru.yandex.money.gradle.plugins.backend.build.JavaModulePlugin;
 import ru.yandex.money.gradle.plugins.gradleproject.publishing.PublishingConfigurer;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.GitExpiredBranchPlugin;
 import ru.yandex.money.gradle.plugins.release.ReleasePlugin;
+import ru.yandex.money.gradle.plugins.task.monitoring.TaskExecutionMonitoringPlugin;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -26,7 +28,8 @@ public class GradleProjectPlugin implements Plugin<Project> {
             MavenPublishPlugin.class,
             ReleasePlugin.class,
             GitExpiredBranchPlugin.class,
-            JavaGradlePluginPlugin.class
+            TaskExecutionMonitoringPlugin.class,
+            ArchitectureTestPlugin.class
     );
 
     @Override
@@ -34,16 +37,17 @@ public class GradleProjectPlugin implements Plugin<Project> {
         if (VersionNumber.parse(project.getGradle().getGradleVersion()).compareTo(VersionNumber.parse("4.10.2'")) < 0) {
             throw new IllegalStateException("Gradle >= 4.10.2 is required");
         }
-        PLUGINS_TO_APPLY.forEach(pluginClass -> project.getPluginManager().apply(pluginClass));
+
+        project.getPluginManager().apply(JavaGradlePluginPlugin.class);
         configureRepos(project);
-        ExtensionConfigurator.configure(project);
+
         new PublishingConfigurer().init(project);
+        PLUGINS_TO_APPLY.forEach(pluginClass -> project.getPluginManager().apply(pluginClass));
+        ExtensionConfigurator.configure(project);
         project.getTasks().create("checkComponentSnapshotDependencies").dependsOn("checkSnapshotsDependencies");
     }
 
     private void configureRepos(Project project) {
         project.getRepositories().maven(repo -> repo.setUrl("https://nexus.yamoney.ru/repository/gradle-plugins/"));
     }
-
-
 }